+++ packages/orq-rc/RELEASES.md
@@ -2869,3 +2869,13 @@ Based on:
 - [typescript v3.10.0-rc.45] packages/orq-rc
 ### Releases
 - [NPM v3.10.0-rc.45] https://www.npmjs.com/package/@orq-ai/node/v/3.10.0-rc.45 - packages/orq-rc
+
+## 2025-08-05 11:13:41
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.596.1 (2.672.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.10.0-rc.46] packages/orq-rc
+### Releases
+- [NPM v3.10.0-rc.46] https://www.npmjs.com/package/@orq-ai/node/v/3.10.0-rc.46 - packages/orq-rc